# Further performance improvement over #675