--- a/oracle/src/lib.rs
+++ b/oracle/src/lib.rs
@@ -9,8 +9,8 @@ mod tests;
 
 use ark_ff::Field;
 
-pub trait FqSponge<Fq: Field, G, Fr> {
-    fn new(p: poseidon::ArithmeticSpongeParams<Fq>) -> Self;
+pub trait FqSponge<'a, Fq: Field, G, Fr> {
+    fn new(p: &'a poseidon::ArithmeticSpongeParams<Fq>) -> Self;
     fn absorb_g(&mut self, g: &[G]);
     fn absorb_fr(&mut self, x: &[Fr]);
     fn challenge(&mut self) -> Fr;