--- a/oracle/src/poseidon.rs
+++ b/oracle/src/poseidon.rs
@@ -8,9 +8,9 @@ use serde_with::serde_as;
 
 /// Cryptographic sponge interface - for hashing an arbitrary amount of
 /// data into one or more field elements
-pub trait Sponge<Input: Field, Digest> {
+pub trait Sponge<'a, Input: Field, Digest> {
     /// Create a new cryptographic sponge using arithmetic sponge `params`
-    fn new(params: ArithmeticSpongeParams<Input>) -> Self;
+    fn new(params: &'a ArithmeticSpongeParams<Input>) -> Self;
 
     /// Absorb an array of field elements `x`
     fn absorb(&mut self, x: &[Input]);
@@ -42,27 +42,27 @@ pub struct ArithmeticSpongeParams<F: Field> {
 }
 
 #[derive(Clone)]
-pub struct ArithmeticSponge<F: Field, SC: SpongeConstants> {
+pub struct ArithmeticSponge<'a, F: Field, SC: SpongeConstants> {
     pub sponge_state: SpongeState,
     rate: usize,
     // TODO(mimoo: an array enforcing the width is better no? or at least an assert somewhere)
     pub state: Vec<F>,
-    params: ArithmeticSpongeParams<F>,
+    params: &'a ArithmeticSpongeParams<F>,
     pub constants: std::marker::PhantomData<SC>,
 }
 
-impl<F: Field, SC: SpongeConstants> ArithmeticSponge<F, SC> {
+impl<'a, F: Field, SC: SpongeConstants> ArithmeticSponge<'a, F, SC> {
     pub fn full_round(&mut self, r: usize) {
-        full_round::<F, SC>(&self.params, &mut self.state, r);
+        full_round::<F, SC>(self.params, &mut self.state, r);
     }
 
     fn poseidon_block_cipher(&mut self) {
-        poseidon_block_cipher::<F, SC>(&self.params, &mut self.state);
+        poseidon_block_cipher::<F, SC>(self.params, &mut self.state);
     }
 }
 
-impl<F: Field, SC: SpongeConstants> Sponge<F, F> for ArithmeticSponge<F, SC> {
-    fn new(params: ArithmeticSpongeParams<F>) -> ArithmeticSponge<F, SC> {
+impl<'a, F: Field, SC: SpongeConstants> Sponge<'a, F, F> for ArithmeticSponge<'a, F, SC> {
+    fn new(params: &'a ArithmeticSpongeParams<F>) -> ArithmeticSponge<F, SC> {
         let capacity = SC::SPONGE_CAPACITY;
         let rate = SC::SPONGE_RATE;