--- a/tools/kimchi-visu/src/main.rs
+++ b/tools/kimchi-visu/src/main.rs
@@ -11,7 +11,7 @@ use mina_curves::pasta::Fp;
 
 fn main() {
     let public = 3;
-    let poseidon_params = oracle::pasta::fp_kimchi::params();
+    let poseidon_params = oracle::pasta::fp_kimchi_params();
 
     // create circuit
     let (gates, row) = {