This PR further improves the performance of creating a hasher or a merger over #675 
However, it introduces more changes to the repo so I make this a separate PR.

Benchmark results on the same machine shows the time cost of `mina_hasher::create_legacy` is further reduced from ~3us to ~60ns (reduced by another ~98%), from originally ~800us (reduced by ~99.9925%)

P.S. If we remove the `reset` mechanism from `Hasher`, the time cost will be reduced from ~60ns to ~20ns (by ~66.6%)

```
legacy/fp_legacy::params()                                                                            
                        time:   [821.75 µs 847.07 µs 878.59 µs]
                        change: [-3.5910% -0.7428% +1.8346%] (p = 0.61 > 0.05)
                        No change in performance detected.
Found 14 outliers among 100 measurements (14.00%)
  5 (5.00%) high mild
  9 (9.00%) high severe
legacy/fp_legacy_params()                                                                            
                        time:   [478.55 ps 482.45 ps 486.86 ps]
                        change: [-6.1538% -2.2139% +1.3907%] (p = 0.28 > 0.05)
                        No change in performance detected.
Found 2 outliers among 100 measurements (2.00%)
  1 (1.00%) high mild
  1 (1.00%) high severe
legacy/create_legacy()  time:   [62.786 ns 63.777 ns 64.860 ns]                                   
                        change: [-7.1318% -5.2283% -3.2964%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 5 outliers among 100 measurements (5.00%)
  4 (4.00%) high mild
  1 (1.00%) high severe

kimchi/fp_kimchi::params()                                                                            
                        time:   [458.25 µs 468.64 µs 482.34 µs]
                        change: [-11.634% -7.2391% -2.9154%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 17 outliers among 100 measurements (17.00%)
  2 (2.00%) high mild
  15 (15.00%) high severe
kimchi/fp_kimchi_params()                                                                            
                        time:   [304.85 ps 311.32 ps 319.88 ps]
                        change: [-5.5933% -1.5054% +2.5485%] (p = 0.49 > 0.05)
                        No change in performance detected.
Found 8 outliers among 100 measurements (8.00%)
  5 (5.00%) high mild
  3 (3.00%) high severe
kimchi/create_kimchi()  time:   [62.066 ns 62.837 ns 64.091 ns]                                   
                        change: [+0.2323% +1.3099% +2.4342%] (p = 0.02 < 0.05)
                        Change within noise threshold.
Found 7 outliers among 100 measurements (7.00%)
  6 (6.00%) high mild
  1 (1.00%) high severe
```